🚚 Move google-firestore functions to a subdirectory
