+++ src/functions/google-firestore/index.ts
@@ -1,0 +1,1 @@
+export { GoogleFirestoreMergeRules } from './merge-rules.js';
